Changed argument to sensible variable object. -Moved definition of sensible variables to its own file -Changed the way all functions passed the information, using the object.
